requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
pydantic==2.5.2
redis==5.0.1
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2
lightgbm==4.1.0
tensorflow-cpu==2.15.0
transformers==4.35.2
torch==2.1.1
tldextract==5.1.1
python-whois==0.8.0
requests==2.31.0
dnspython==2.4.2
sqlalchemy==2.0.23
google-generativeai==0.3.1
python-multipart==0.0.6
psycopg2-binary==2.9.9
joblib==1.3.2
huggingface-hub==0.19.4
python-dotenv==1.0.0
prometheus-fastapi-instrumentator==6.1.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-redis==0.42b0
cryptography==41.0.5
pyOpenSSL==23.3.0
httpx==0.25.2
